// packages/backend/src/utils/types/error.types.ts
export enum AuthErrorTypes {
    USER_EXISTS = 'USER_EXISTS',
    REGISTRATION_FAIL = 'REGISTRATION_FAIL',
    INVALID_UNI_EMAIL = 'INVALID_UNI_EMAIL',
    INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',
    MISSING_CREDENTIALS = 'MISSING_CREDENTIALS',
    NOT_AUTHENTICATED = 'NOT_AUTHENTICATED',
    MISSING_TOKEN = 'MISSING_TOKEN',
}